Fix TypeError in QLoggingTableModel for py38
py38 tests are failing due to an stricter type checking in QObject.startTimer()
Cast float argument to int to fix it.

+++ lib/taurus/qt/qtgui/table/qlogtable.py
@@ -156,7 +156,7 @@ class QLoggingTableModel(Qt.QAbstractTableModel, logging.Handler):
         self._records = []
         self._accumulated_records = []
         Logger.addRootLogHandler(self)
-        self.startTimer(freq * 1000)
+        self.startTimer(int(freq * 1000))
 
     # ---------------------------------
     # Qt.QAbstractTableModel overwrite
